requirements.txt
pandas
matplotlib
flask
requests
flask-sqlalchemy
psycopg2